updated test data

+++ models/utils.go
@@ -2,6 +2,6 @@ package models
 
 import "math"
 
-func RoundBells(bells float64) int {
-	return int(math.Ceil(bells))
+func RoundBells(bells float32) int {
+	return int(math.Ceil(float64(bells)))
 }
